refactor(tap): colocate React memo-cache fallbacks; dissolve core memo-cache helper
reactC and useReactMemoCache live in react-shim/useReactMemoCache.ts next
to the shim they fall back to; the memo-cache primitives move into
react-hooks/useMemoCache.ts now that nothing in core needs them.
