+++ packages/tap/src/react-shim/index.ts
@@ -16,7 +16,7 @@ import {
   isReadableTapContext,
 } from "../core/context";
 import { useReactEffectEvent } from "./useReactEffectEvent";
-import { useReactMemoCacheShim } from "./useReactMemoCache";
+import { useReactMemoCache } from "./useReactMemoCache";
 
 // @ts-expect-error -- @types/react uses `export =`; this is valid at runtime.
 export * from "react";
@@ -71,8 +71,6 @@ export const useSyncExternalStore = (
         getSnapshot,
         getServerSnapshot,
       );
-
-const useReactMemoCache = ReactRuntime.useMemoCache ?? useReactMemoCacheShim;
 
 export const useMemoCache = (size: number) =>
   inTap() ? hooks.useMemoCache(size) : useReactMemoCache(size);
