+++ packages/tap/src/react-shim/useReactMemoCache.ts
@@ -1,9 +1,11 @@
 /* oxlint-disable react/exhaustive-deps -- once-per-mount memo cell */
-import { useMemo } from "react";
+import React, { useMemo } from "react";
 import {
   MEMO_CACHE_SENTINEL,
   createMemoCache,
-} from "../core/helpers/memo-cache";
+} from "../react-hooks/useMemoCache";
+
+const ReactRuntime = React as any;
 
 // Mirror of Meta's compiler-runtime polyfill for React builds without a
 // native memo cache: a once-per-mount memo cell.
@@ -14,3 +16,9 @@ export const useReactMemoCacheShim = (size: number): unknown[] =>
     ($ as any)[MEMO_CACHE_SENTINEL] = true;
     return $;
   }, []);
+
+export const reactC: (size: number) => unknown[] =
+  ReactRuntime.__COMPILER_RUNTIME?.c ?? useReactMemoCacheShim;
+
+export const useReactMemoCache: (size: number) => unknown[] =
+  ReactRuntime.useMemoCache ?? useReactMemoCacheShim;
